backend/requirements.txt
fastapi
uvicorn
edge-tts
cloudinary
pymongo
apscheduler
httpx
python-multipart
python-dotenv
